function-rotate-secret/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets